Tips Collected at a restaurant
The data set is from a restaurant from a city that operates only on 4 days of the week. Thursday, Friday, Saturday, Sunday. It records each transaction that was done when the restaurant was open.

The key parameters are

Total Bill - Amount of total bill in USD
Tip - Amount of tip given by the customer in USD
Sex - Sax of the guest who pays for the bill (Male or Female)
Smoker - Smoker in the party (Yes or No)
Day - Day of the week (Thu, Fri, Sat, Sun)
Time - Dinner or Lunch
Size - Number of the guests in the party
As a data scientist, you have been assigned the task to investigate the pattern of the tip received by the restaurant staff because the tip is an important parameter of their salary

Find out Five-Number summary for "Total Bill" and "Tip"
Plot the Box plot diagram for "Total Bill" and "Tip"
Analyze the diagram to mark the skewness in the data
Find out the outliers for "Total Bill" and "Tip"
Find out the IQR
Plot the histogram for "Total Bill" and "Tip"
Identify skewness in each. Type of skewness and the possible reason for the skewness
Plot the cumulative Frequency Polygon for "Total Bill" and "Tip"
Prepare the Frequency Table and Bar Chart for "Size". Summarize and explain your findings
Prepare two variable frequency tables for "Size" vs "Tip" and "Size" vs "Total Bill". Summarize and explain your findings.
Explore if there is any dependency between the variable "Tip" and rest of the variables