Sales of Summer Clothes
The Data Set: Sales Data of Summer Clothes from E-Commerce Company

Explore the sales data of an e-commerce company. The data set has the product rating, prices of various products, and sales performance along with other parameters.

As a data scientist, you have been assigned a task of analyzing various parameters that effects the sales.

Conduct an Exploratory Data Analysis to find the following

Load the csv file to pandas
Check which columns are present
Check and drop duplicate values
Check the null values (When you execute sp.info(), you can see that various columns have lesser number of records than others that shows the presence of null values)
Count the number of null values in each column. Analyze the possible reasons for the missing value and document the findings if any
Calculate count, mean, std, min, 25%, 50%, 75%, max values for each column. Prepare an analysis of the difference between mean and median for each column and possible reasons for the same. 
Prepare Histograms and Countplots
Boxplot of ratings