Study the data set of white wine quality.
The objective of the study is to conduct EDA on several parameters of the data set. 

Exploratory Data Analysis

Understand the data using the pandas library
Print the first five rows of the dataset
Print the last five rows of the dataset
Find out the total number of rows and columns of the data set
Find out the columns, data types, and presence of null values or missing values in the data set
Univariate Analysis
Find out missing values graphically
Find out the number of rows, mean, std deviation, min, Q1, Q2, Q3, max values for each variable. Document your observations to check if the outliers are present in the data set
Draw the histogram, kernel density estimate (kde) to check the distribution and skewness for each variable. Document your observation
Create a frequency distribution table and bar chart for the output variable
Draw box plot for each variable and identify the IQR and outliers
Multivariate Analysis
Draw the correlation matrix and identify the variables that are correlated to each other.